requirements.txt
sudo apt-get install git
Django==2.2.2
djangorestframework==3.9.4
psycopg2==2.8.3
pytz==2019.1
sqlparse==0.3.0
